Creating jwt_svid_validator
Extracting a class from JWT-Svid to make it eaiser to test and read the code.
