+++ src/pyspiffe/svid/jwt_svid.py
@@ -1,21 +1,11 @@
-import datetime
-from calendar import timegm
 import jwt
+import datetime
 from pyspiffe.spiffe_id.spiffe_id import SpiffeId
 from pyspiffe.bundle.jwt_bundle.jwt_bundle import JwtBundle
-from pyspiffe.svid.exceptions import (
-    TokenExpiredError,
-    InvalidClaimError,
-    InvalidAlgorithmError,
-    InvalidTypeError,
-)
-import datetime
-from calendar import timegm
-import jwt
-from jwt import exceptions
+from pyspiffe.svid.jwt_svid_validator import JwtSvidValidator
+
 
 INVALID_INPUT_ERROR = 'Invalid input: {}.'
-AUDIENCE_NOT_MATCH_ERROR = 'audience does not match expected value'
 
 
 class JwtSvid(object):
@@ -24,23 +14,9 @@ class JwtSvid(object):
     See <a href="https://github.com/spiffe/spiffe/blob/master/standards/JWT-SVID.md">https://github.com/spiffe/spiffe/blob/master/standards/JWT-SVID.md</a>.
     """
 
-    _required_claims = ['aud', 'exp', 'sub']
-    _supported_algorithms = [
-        'RS256',
-        'RS384',
-        'RS512',
-        'ES256',
-        'ES384',
-        'ES512',
-        'PS256',
-        'PS384',
-        'PS512',
-    ]
-    _supported_types = ['JWT', 'JOSE']
-
     def __init__(
         self, spiffeId: SpiffeId, audience: [], expiry: datetime, claims: {}, token: str
-    ):
+    ) -> None:
         self.spiffeId = spiffeId
         self.audience = audience
         self.expiry = expiry
@@ -74,9 +50,10 @@ class JwtSvid(object):
         if not token:
             raise ValueError(INVALID_INPUT_ERROR.format('token cannot be empty'))
         token_header = jwt.get_unverified_header(token)
-        cls._validate_header(token_header)
+        validator = JwtSvidValidator()
+        validator.validate_header(token_header)
         claims = jwt.decode(token, options={'verify_signature': False})
-        cls._validate_claims(claims, expected_audience)
+        validator.validate_claims(claims, expected_audience)
         spiffe_ID = SpiffeId.parse(claims['sub'])
         result = JwtSvid(spiffe_ID, claims['aud'], claims['exp'], claims, token)
 
@@ -136,75 +113,3 @@ class JwtSvid(object):
         result = JwtSvid(spiffe_ID, claims['aud'], claims['exp'], claims, token)
 
         return result
-
-    ###
-    ### Verifies if header specifies supported algortihms and token type.
-    ### Raises:
-    ###     ValueError: in case header is not specified.
-    ###     InvalidAlgorithmError: in case specified 'alg' is not supported as specified by the SPIFFE standard.
-    ###     InvalidTypeError: in case 'typ' is present in header but is not set to 'JWT' or 'JOSE'.
-    ###
-    @classmethod
-    def _validate_header(cls, header: {}) -> None:
-        if not header:
-            raise ValueError(INVALID_INPUT_ERROR.format('header cannot be empty'))
-
-        if header['alg'] not in cls._supported_algorithms:
-            raise InvalidAlgorithmError(header['alg'])
-        try:
-            if header['typ'] and header['typ'] not in cls._supported_types:
-                raise InvalidTypeError(header['typ'])
-        except KeyError:
-            pass
-
-    ###
-    ### Verifies token expiration.
-    ### Note: If and when https://github.com/jpadilla/pyjwt/issues/599 is fixed, this can be simplified/removed.
-    ### Raises:
-    ###     TokenExpiredError: in case token is expired.
-    ###
-    @classmethod
-    def _validate_exp(cls, expiration_date: str) -> None:
-        expiration_date = int(expiration_date)
-        utctime = timegm(datetime.datetime.utcnow().utctimetuple())
-        if expiration_date < utctime:
-            raise TokenExpiredError()
-
-    ###
-    ### Verifies if expected_audience is present in audience_claim. The aud claim MUST be present.
-    ### Raises:
-    ###     InvalidClaimError: in expected_audience is not a subset of audience_claim.
-    ###     ValueError: in case expected_audience is empty.
-    ###
-    @classmethod
-    def _validate_aud(cls, audience_claim: [], expected_audience: []) -> None:
-        if not expected_audience:
-            raise ValueError(
-                INVALID_INPUT_ERROR.format('expected_audience cannot be empty')
-            )
-
-        if not audience_claim or all(aud == '' for aud in audience_claim):
-            raise InvalidClaimError('audience_claim cannot be empty')
-
-        if not all(aud in audience_claim for aud in expected_audience):
-            raise InvalidClaimError(AUDIENCE_NOT_MATCH_ERROR)
-
-    ###
-    ### Validates payload for required claims (aud, exp, sub) - signature expiration and audience.
-    ###
-    ### Raises:
-    ###     InvalidClaimError: in case a required claim is not present in payload or expected_audience is not a subset of audience_claim.
-    ###     TokenExpiredError: in case token is expired.
-    ###     ValueError: in case expected_audience is empty.
-
-    ###
-    @classmethod
-    def _validate_claims(cls, payload: {}, expected_audience: []) -> None:
-        try:
-            for claim in cls._required_claims:
-                if not payload[claim]:
-                    raise InvalidClaimError(claim)
-            cls._validate_exp(payload['exp'])
-            cls._validate_aud(payload['aud'], expected_audience)
-        except KeyError as key_error:
-            raise InvalidClaimError(key_error.args[0])
